multi-source-econ-pipeline/requirements.txt
dbt-snowflake==1.11.3
snowflake-connector-python==4.4.0
requests
pandas
python-dotenv